Initial Commit
Created a library to serialize and deserialize bencode using serde. I'm
not sure how I want to handle the plethora of unimplemented functions.
Not sure if it is alright to just keep them unimplemented or if I should
make it more explicit in the code and error messages that it is
purposefully unimplemented.

Probably need a better implementation of src/error.rs keeping track of
line numbers and actual error messages and what not. Need to standardize
how the code should respond to malformed entries and get those unit
tested as well. Need to properly set up docstrings to look as
professional as serde_json. Need to find a way to gate the serde derive
feature behind testing.

Basic implementation and Unit Tests for Serialization (ser)
Basic implementation and Unit Tests for Deserialization (de)