Generic Read Template
Trying to use the concept from serde_json of using a Read trait to create generic deserializer's for the different way's
of consuming data to deserialize. This is mostly to allow deserializing from bytes which is neccessary for Binary encodings like bencode as well as
to keep tests simple that just want to use proper UTF-8 Strings.